requirements.txt
amqp==5.1.1
arabic-reshaper==3.0.0
asgiref==3.5.2
asn1crypto==1.5.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.29.0
cryptography==37.0.4
cssselect2==0.6.0
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.2
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-dotenv==1.4.2
django-filter==21.1
django-mathfilters==1.0.0
django-model-utils==4.2.0
django-timezone-field==5.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
docopt==0.6.2
exceptiongroup==1.1.0
fontawesome-free==5.15.4
fontawesomefree==6.1.1
future==0.18.2
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
idna==3.3
kombu==5.2.4
lxml==4.9.1
num2words==0.5.12
numpy==1.23.0
oscrypto==1.3.0
outcome==1.2.0
packaging==21.3
pandas==1.5.0
pdfkit==1.0.0
Pillow==9.1.1
prompt-toolkit==3.0.30
psycopg==3.1.9
pycparser==2.21
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
pyparsing==3.0.9
PyPDF3==1.0.6
PySocks==1.7.1
python-bidi==0.4.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
rabbitmq==0.2.0
redis==4.3.4
reportlab==3.6.11
requests==2.28.0
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stripe==5.2.0
svglib==1.4.1
swapper==1.3.0
tinycss2==1.1.1
tqdm==4.64.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.2.0
wrapt==1.14.1
wsproto==1.2.0
xhtml2pdf==0.2.8
